Moved ClientContainer to shared.
