+++ Robust.Shared/Containers/Container.cs
@@ -141,6 +141,7 @@ namespace Robust.Shared.Containers
             DebugTools.Assert(!Deleted);
 
             Owner.EntityManager.EventBus.RaiseEvent(EventSource.Local, new EntInsertedIntoContainerMessage(toinsert, this));
+            Owner.EntityManager.EventBus.RaiseEvent(EventSource.Local, new UpdateContainerOcclusionMessage(toinsert));
             Manager.Owner.SendMessage(Manager, new ContainerContentsModifiedMessage(this, toinsert, false));
             Manager.Dirty();
         }
@@ -201,7 +202,7 @@ namespace Robust.Shared.Containers
             DebugTools.Assert(toremove.IsValid());
 
             Owner?.EntityManager.EventBus.RaiseEvent(EventSource.Local, new EntRemovedFromContainerMessage(toremove, this));
-
+            Owner?.EntityManager.EventBus.RaiseEvent(EventSource.Local, new UpdateContainerOcclusionMessage(toremove));
             Manager.Owner.SendMessage(Manager, new ContainerContentsModifiedMessage(this, toremove, true));
             Manager.Dirty();
         }
